+++ test/src/main/java/org/vertexium/test/GraphTestBase.java
@@ -1823,6 +1823,25 @@ public abstract class GraphTestBase {
 
         vertices = graph.query(AUTHORIZATIONS_A).has("notSetProp", Compare.LESS_THAN_EQUAL, 5).vertices();
         assertResultsCount(0, 0, vertices);
+
+        vertices = graph.query(AUTHORIZATIONS_A).hasId("v1").vertices();
+        assertResultsCount(1, 1, vertices);
+
+        vertices = graph.query(AUTHORIZATIONS_A).hasId("v1", "v2").vertices();
+        assertResultsCount(2, 2, vertices);
+
+        vertices = graph.query(AUTHORIZATIONS_A).hasId("v1", "v2").hasId("v1").vertices();
+        assertResultsCount(1, 1, vertices);
+        assertVertexIds(vertices, new String[] {"v1"});
+
+        vertices = graph.query(AUTHORIZATIONS_A).hasId("v1").hasId("v2").vertices();
+        assertResultsCount(0, 0, vertices);
+
+        edges = graph.query(AUTHORIZATIONS_A).hasId("e1").edges();
+        assertResultsCount(1, 1, edges);
+
+        edges = graph.query(AUTHORIZATIONS_A).hasId("e1", "e2").edges();
+        assertResultsCount(2, 2, edges);
     }
 
     @Test